Move legacy tracker to its own package
