+++ src/meltano/cli/install.py
@@ -1,8 +1,9 @@
 import click
+
+from meltano.core.legacy_tracking import LegacyTracker
 from meltano.core.plugin import PluginType
 from meltano.core.project_add_service import ProjectAddService
 from meltano.core.project_plugins_service import ProjectPluginsService
-from meltano.core.tracking import LegacyTracker
 
 from . import cli
 from .params import pass_project
