+++ src/meltano/cli/select.py
@@ -5,10 +5,10 @@ from typing import Dict
 import click
 
 from meltano.core.db import project_engine
+from meltano.core.legacy_tracking import LegacyTracker
 from meltano.core.plugin.error import PluginExecutionError
 from meltano.core.plugin.singer.catalog import SelectionType, SelectPattern
 from meltano.core.select_service import SelectService
-from meltano.core.tracking import LegacyTracker
 from meltano.core.utils import click_run_async
 
 from . import cli
