+++ src/meltano/cli/validate.py
@@ -10,8 +10,8 @@ from sqlalchemy.orm.session import sessionmaker
 
 from meltano.cli.utils import propagate_stop_signals
 from meltano.core.db import project_engine
+from meltano.core.legacy_tracking import LegacyTracker
 from meltano.core.project import Project
-from meltano.core.tracking import LegacyTracker
 from meltano.core.utils import run_async
 from meltano.core.validation_service import ValidationOutcome, ValidationsRunner
