+++ src/meltano/core/tracking/__init__.py
@@ -1,4 +1,3 @@
 """Meltano telemetry."""
 
-from .legacy_tracker import LegacyTracker
 from .tracker import Tracker
